Audio/Mixer: Got rid of the stack-allocated scratch buffer when mixing / multiplexing stuff. Instead, each sink has a heap allocated one now, to avoid hammering the stack [build fix]. ticketoem2ref:36
svn:sync-xref-src-repo-rev: r142232

+++ src/VBox/Devices/Audio/AudioMixer.cpp
@@ -1,4 +1,4 @@
-/* $Id: AudioMixer.cpp 87270 2021-01-15 13:06:46Z andreas.loeffler@oracle.com $ */
+/* $Id: AudioMixer.cpp 87273 2021-01-15 13:37:58Z andreas.loeffler@oracle.com $ */
 /** @file
  * Audio mixing routines for multiplexing audio sources in device emulations.
  *
@@ -1711,7 +1711,7 @@ static int audioMixerSinkUpdateInternal(PAUDMIXSINK pSink)
     while (cbToWriteToStreams)
     {
         uint32_t cfChunk;
-        rc  = AudioMixBufAcquireReadBlock(&pSink->MixBuf, pSink->pabScratchBuf, RT_MIN(cbToWriteToStreams, pSink->cbScratchBuf),
+        rc  = AudioMixBufAcquireReadBlock(&pSink->MixBuf, pSink->pabScratchBuf, RT_MIN(cbToWriteToStreams, (uint32_t)pSink->cbScratchBuf),
                                           &cfChunk);
         if (RT_FAILURE(rc))
             break;
